pyproject.toml
[tool.poetry]
name = "websec-2"
version = "0.1.0"
description = ""
authors = ["Ilya Zazvonov <izazvonov@rvision.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.135.1"
beautifulsoup4 = "^4.14.3"
pydantic-settings = "^2.13.1"
redis = "^7.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
